implemented maximized/monocle windows

+++ src/ewmh.c
@@ -37,7 +37,7 @@ void ewmh_create_ewmh_window(xcb_connection_t * conn, int scrno) {
         XCB_CW_OVERRIDE_REDIRECT,
         (uint32_t[]){1}
     );
-    
+
     ( void ) xcb_ewmh_set_wm_name(ewmh, ewmh_w, WMNAMELEN, WMNAME);
     ( void ) xcb_ewmh_set_supporting_wm_check(ewmh, root_screen->root, ewmh_w);
 
@@ -45,7 +45,7 @@ void ewmh_create_ewmh_window(xcb_connection_t * conn, int scrno) {
     ( void ) xcb_ewmh_set_wm_pid(ewmh, root_screen->root, getpid());
     ( void ) xcb_ewmh_set_supporting_wm_check(ewmh, root_screen->root, root_screen->root);
 
-    
+
     xcb_atom_t supported[] = {
         #define xmacro(ATOM) ewmh->ATOM,
             #include "atoms.xmacro"
@@ -53,31 +53,31 @@ void ewmh_create_ewmh_window(xcb_connection_t * conn, int scrno) {
     };
 
     ( void ) xcb_ewmh_set_supported(
-        ewmh, scrno, 
+        ewmh, scrno,
         sizeof(supported) / sizeof(*supported),
         supported
     );
 
 
     ( void ) xcb_map_window(conn, ewmh_w);
-    ( void ) xcb_configure_window(conn, ewmh_w, XCB_CONFIG_WINDOW_STACK_MODE, (uint32_t[]){XCB_STACK_MODE_BELOW});   
+    ( void ) xcb_configure_window(conn, ewmh_w, XCB_CONFIG_WINDOW_STACK_MODE, (uint32_t[]){XCB_STACK_MODE_BELOW});
 
     xcb_flush(conn);
 }
 
-void ewmh_change_desktop_number(uint8_t scrno, uint8_t desktop_no){
+void ewmh_change_desktop_number(uint8_t scrno, uint32_t desktop_no){
     ( void ) xcb_ewmh_set_number_of_desktops(ewmh, scrno, desktop_no);
 }
 
 
-void ewmh_change_to_desktop(int scrno, uint8_t desktop){
+void ewmh_change_to_desktop(int scrno, uint32_t desktop){
     if (desktop >= MAX_DESKTOPS) return;
 
     ( void ) xcb_ewmh_set_current_desktop(ewmh, scrno, desktop);
 }
 
 
-void ewmh_create_desktops(int scrno, uint8_t desktop_no) {
+void ewmh_create_desktops(int scrno, uint32_t desktop_no) {
     ewmh_change_desktop_number(scrno, desktop_no);
     ewmh_change_to_desktop(scrno, 0);
 }
